requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.5
python-dotenv==1.0.0
werkzeug==3.0.1
Jinja2==3.1.2

